Move FasterCPythonResults to _pyperformance.
