+++ PORTAL/jobs/requests.py
@@ -97,6 +97,9 @@ class RequestID(namedtuple('RequestID', 'kind timestamp user')):
         return dt
 
 
+##################################
+# local metadata
+
 class Request(_utils.Metadata):
 
     FIELDS = [
@@ -886,134 +889,3 @@ def build_compile_script(req, bfiles, fake=None):
 
         #rm -f {bfiles.work.pidfile}
     '''[1:-1])
-
-
-##################################
-# pyperformance helpers
-
-class PyperfResultsStorage:
-
-    def add(self, results, *, unzipped=True):
-        raise NotImplementedError
-
-
-class PyperfResultsRepo(PyperfResultsStorage):
-
-    BRANCH = 'add-benchmark-results'
-
-    def __init__(self, root, remote=None, datadir=None):
-        if root:
-            root = os.path.abspath(root)
-        if remote:
-            if isinstance(remote, str):
-                remote = _utils.GitHubTarget.resolve(remote, root)
-            elif not isinstance(remote, _utils.GitHubTarget):
-                raise TypeError(f'unsupported remote {remote!r}')
-            root = remote.ensure_local(root)
-        else:
-            if root:
-                if not os.path.exists(root):
-                    raise FileNotFoundError(root)
-                #_utils.verify_git_repo(root)
-            else:
-                raise ValueError('missing root')
-            remote = None
-        self.root = root
-        self.remote = remote
-        self.datadir = datadir or None
-
-    def git(self, *args, cfg=None):
-        ec, text = _utils.git(*args, cwd=self.root, cfg=cfg)
-        if ec:
-            raise NotImplementedError((ec, text))
-        return text
-
-    def add(self, results, *,
-            branch=None,
-            author=None,
-            unzipped=True,
-            split=True,
-            push=True,
-            ):
-        if not branch:
-            branch = self.BRANCH
-
-        if not isinstance(results, _pyperformance.PyperfResults):
-            raise NotImplementedError(results)
-        source = results.filename
-        if not os.path.exists(source):
-            logger.error(f'results not found at {source}')
-            return
-        if split:
-            by_suite = results.split_benchmarks()
-            if 'other' in by_suite:
-                raise NotImplementedError(sorted(by_suite))
-        else:
-            by_suite = {None: results}
-
-        if self.remote:
-            self.remote.ensure_local(self.root)
-
-        authorargs = ()
-        cfg = {}
-        if not author:
-            pass
-        elif isinstance(author, str):
-            parsed = _utils.parse_email_address(author)
-            if not parsed:
-                raise ValueError(f'invalid author {author!r}')
-            name, email = parsed
-            if not name:
-                name = '???'
-                author = f'??? <{author}>'
-            cfg['user.name'] = name
-            cfg['user.email'] = email
-            #authorargs = ('--author', author)
-        else:
-            raise NotImplementedError(author)
-
-        logger.info(f'adding results {source}...')
-        for suite in by_suite:
-            suite_results = by_suite[suite]
-            name = suite_results.get_upload_name()
-            reltarget = f'{name}.json'
-            if self.datadir:
-                reltarget = f'{self.datadir}/{reltarget}'
-            if not unzipped:
-                reltarget += '.gz'
-            target = os.path.join(self.root, reltarget)
-
-            logger.info(f'...as {target}...')
-            self.git('checkout', '-B', branch)
-            if unzipped:
-                data = suite_results.data
-                with open(target, 'w') as outfile:
-                    json.dump(data, outfile, indent=2)
-            else:
-                shutil.copyfile(source, target)
-            self.git('add', reltarget)
-            msg = f'Add Benchmark Results ({name})'
-            self.git('commit', *authorargs, '-m', msg, cfg=cfg)
-        logger.info('...done adding')
-
-        if push:
-            self._upload(reltarget)
-
-    def _upload(self, reltarget):
-        if not self.remote:
-            raise Exception('missing remote')
-        url = f'{self.remote.url}/tree/main/{reltarget}'
-        logger.info(f'uploading results to {url}...')
-        self.git('push', self.remote.push_url)
-        logger.info('...done uploading')
-
-
-class FasterCPythonResults(PyperfResultsRepo):
-
-    REMOTE = _utils.GitHubTarget.from_origin('faster-cpython', 'ideas', ssh=True)
-    DATADIR = 'benchmark-results'
-
-    def __init__(self, root=None, remote=None):
-        if not remote:
-            remote = self.REMOTE
-        super().__init__(root, remote, self.DATADIR)
